requirements.txt
geoip2==4.2.0
requests==2.27.1
ruamel.base==1.0.0
scapy==2.4.5
